+++ autorag/nodes/passagefilter/__init__.py
@@ -1,4 +1,5 @@
-from .percentile_cutoff import similarity_percentile_cutoff
 from .pass_passage_filter import pass_passage_filter
 from .recency import recency_filter
-from .threshold_cutoff import similarity_threshold_cutoff
+from .similarity_percentile_cutoff import similarity_percentile_cutoff
+from .similarity_threshold_cutoff import similarity_threshold_cutoff
+from .threshold_cutoff import threshold_cutoff
