+++ docs/source/nodes/passage_filter/passage_filter.md
@@ -52,4 +52,5 @@ maxdepth: 1
 similarity_threshold_cutoff.md
 similarity_percentile_cutoff.md
 recency_filter.md
+threshold_cutoff.md
 ```
