+++ docs/source/nodes/passage_filter/threshold_cutoff.md
@@ -1,0 +1,22 @@
+# Threshold Cutoff
+
+This module is inspired by
+our [similarity threshold cutoff](https://marker-inc-korea.github.io/AutoRAG/nodes/passage_filter/similarity_threshold_cutoff.html)
+module.
+Filters the contents, scores, and ids based on a **previous result's scores**.
+
+📣 Keeps at least one item per query if all scores are below the threshold.
+
+## **Module Parameters**
+
+- **threshold** : The threshold value to filter out the contents.
+  If the similarity score is below the threshold, the content will be filtered out.
+  This is essential to run the module, so you have to set this parameter.
+
+## **Example config.yaml**
+
+```yaml
+modules:
+  - module_type: threshold_cutoff
+    threshold: 0.85
+```
