+++ sample_config/full.yaml
@@ -87,8 +87,12 @@ node_lines:
         - module_type: pass_passage_filter
         - module_type: similarity_threshold_cutoff
           threshold: 0.85
+        - module_type: similarity_percentile_cutoff
+          threshold: 0.6
         - module_type: recency_filter
           threshold: 2015-01-01
+        - module_type: threshold_cutoff
+          threshold: 0.85
     - node_type: passage_compressor
       strategy:
         metrics: [retrieval_token_f1, retrieval_token_recall, retrieval_token_precision]
